requirements.txt
# Deep Learning
tensorflow>=2.16.0
keras>=3.0.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Configuration
PyYAML==6.0.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Web Driver
webdriver-manager==4.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jupyter (optional)
jupyter==1.0.0
ipykernel==6.27.1

# Additional
scikit-learn==1.3.2
joblib==1.3.2